\name{RFC.date-class}
\Rdversion{1.1}
\docType{class}
\alias{RFC.date-class}
\alias{coerce,character,RFC.date-method}

\title{Class "RFC.date"}
\description{
Data to coerce upon input
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{coerce}{\code{signature(from = "character", to = "RFC.date")}: ... }
	 }
}