\name{RFC.list-class}
\Rdversion{1.1}
\docType{class}
\alias{RFC.list-class}
\alias{coerce,character,RFC.list-method}

\title{Class "RFC.list"}
\description{
Data to coerce upon input
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{coerce}{\code{signature(from = "character", to = "RFC.list")}: ... }
	 }
}